requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
colorama==0.4.6
dj-database-url==1.2.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
